// scripts/datagen/pokemon-raw.type.ts
export interface PokemonRaw {
  abilities: Ability[];
  baseExperience: number;
  forms: Species[];
  gameIndices: GameIndex[];
  height: number;
  // heldItems: any[];
  id: number;
  isDefault: boolean;
  locationAreaEncounters: string;
  moves: Move[];
  name: string;
  order: number;
  // pastAbilities: any[];
  // pastTypes: any[];
  species: Species;
  sprites: Sprites;
  stats: Stat[];
  types: Type[];
  weight: number;
}

interface Ability {
  ability: Species;
  isHidden: boolean;
  slot: number;
}

interface Species {
  name: string;
  url: string;
}

interface GameIndex {
  gameIndex: number;
  version: Species;
}

interface Move {
  move: Species;
  versionGroupDetails: VersionGroupDetail[];
}

interface VersionGroupDetail {
  levelLearnedAt: number;
  moveLearnMethod: MoveLearnMethod;
  versionGroup: VersionGroup;
}
// ...
interface Sprites {
  backDefault: string;
  backFemale?: string;
  backShiny: string;
  backShinyFemale?: string;
  frontDefault: string;
  frontFemale?: string;
  frontShiny: string;
  frontShinyFemale?: string;
  other: SpritesOther;
  versions: SpritesVersions;
}

interface SpritesOther {
  dreamWorld: SpritesDreamWorld;
  home: SpritesHome;
  'official-artwork': SpritesOfficialArtwork;
  showdown: SpritesShowdown;
}

interface SpritesDreamWorld {
  frontDefault: string;
  frontFemale?: string;
}

interface SpritesHome {
  frontDefault: string;
  frontFemale?: string;
  frontShiny: string;
  frontShinyFemale?: string;
}

interface SpritesOfficialArtwork {
  frontDefault: string;
  frontShiny: string;
}

type SpritesShowdown = Omit<Sprites, 'other' | 'versions'>;

interface SpritesVersions {
  'generation-i'?: SpritesGenerationI;
  'generation-ii'?: SpritesGenerationII;
  'generation-iii'?: SpritesGenerationIII;
  'generation-iv'?: SpritesGenerationIV;
  'generation-v'?: SpritesGenerationV;
  'generation-vi'?: SpritesGenerationVI;
  'generation-vii'?: SpritesGenerationVII;
  'generation-viii'?: SpritesGenerationVIII;
}
// ...
interface Stat {
  baseStat: number;
  effort: number;
  stat: Species;
}

interface Type {
  slot: number;
  type: Species;
}